requirements.txt
streamlit
torch
torchvision
numpy
opencv-python
lime
shap
matplotlib
scikit-image
Pillow
reportlab